Simplify cache implementation and usage
In practice the only supported operation mode is using an actual cache
directory, so there is not much of a point of keeping a trait plus two
implementations of it, where one is never used.

This removes DummyCache, the Cache trait, and renames XdgCache as Cache.
Also Cache.has() is only used for tests, therefore a #[cfg(test)] guard
was added to it, and may be removed later if needed.

Closes #4 as there is no longer a fallback cache mechanism.