Calculate checksum over the whole generated tarball
This shifts the responsibility of checksum calculation from
bindep::Solver to a separate csum::CSumWriter which implements
std::io::Write, keeps a Blake2b updated and relays writes to
something else which implements Write.

The bindep::Solver gets passed a CSumWriter, so all that it writes
to the Tar file (headers, attributes, etc. and not only file contents)
ends up being part of the checksum.

Also, this writes the checksum result to the cache file directly as
bytes, without converting them first to an hex string.

Fixes #1